+++ src/main/java/com/zincoid/nullbot/core/context/BotCtx.java
@@ -3,14 +3,14 @@ package com.zincoid.nullbot.core.context;
 import com.mikuac.shiro.core.Bot;
 import com.mikuac.shiro.dto.event.Event;
 import com.zincoid.nullbot.core.component.ai.chat.enums.ChatScope;
+import com.zincoid.nullbot.core.enums.EventScope;
 import com.zincoid.nullbot.core.model.data.po.SettingPO;
 
 public final class BotCtx {
 
     private static final ThreadLocal<Bot> bot = new ThreadLocal<>();
     private static final ThreadLocal<Event> event = new ThreadLocal<>();
-
-    private static final ThreadLocal<Boolean> isPrivate = new ThreadLocal<>();
+    private static final ThreadLocal<EventScope> scope = new ThreadLocal<>();
     private static final ThreadLocal<Long> userId = new ThreadLocal<>();
     private static final ThreadLocal<Long> groupId = new ThreadLocal<>();
     private static final ThreadLocal<SettingPO> setting = new ThreadLocal<>();
@@ -19,13 +19,10 @@ public final class BotCtx {
 
     // =================== 系统资源方法 ===================
 
-    public static void setBot(Bot bot) {
+    public static void setCore(Bot bot, Event event) {
         BotCtx.bot.set(bot);
-    }
-    public static void setEvent(Event event) {
         BotCtx.event.set(event);
     }
-
     public static Bot getBot() {
         return bot.get();
     }
@@ -36,31 +33,26 @@ public final class BotCtx {
     // =================== 基本数据方法 ===================
 
     public static void setGroup(Long userId, Long groupId, SettingPO setting) {
-        setIsPrivate(false);
-        setUserId(userId);
-        setGroupId(groupId);
-        setSetting(setting);
-    }
-    public static void setPrivate(Long userId) {
-        setIsPrivate(true);
-        setUserId(userId);
-    }
-
-    public static void setIsPrivate(boolean isPrivate) {
-        BotCtx.isPrivate.set(isPrivate);
-    }
-    public static void setUserId(Long userId) {
+        BotCtx.scope.set(EventScope.GROUP);
         BotCtx.userId.set(userId);
-    }
-    public static void setGroupId(Long groupId) {
         BotCtx.groupId.set(groupId);
-    }
-    public static void setSetting(SettingPO setting) {
         BotCtx.setting.set(setting);
     }
+    public static void setPrivate(Long userId) {
+        BotCtx.scope.set(EventScope.PRIVATE);
+        BotCtx.userId.set(userId);
+        BotCtx.groupId.set(0L);
+        BotCtx.setting.remove();
+    }
+    public static void setUnknown() {
+        BotCtx.scope.set(EventScope.UNKNOWN);
+        BotCtx.userId.remove();
+        BotCtx.groupId.remove();
+        BotCtx.setting.remove();
+    }
 
-    public static boolean getIsPrivate() {
-        return isPrivate.get();
+    public static EventScope getScope() {
+        return scope.get();
     }
     public static Long getUserId() {
         return userId.get();
@@ -77,18 +69,18 @@ public final class BotCtx {
     public static void remove() {
         bot.remove();
         event.remove();
-        isPrivate.remove();
+        scope.remove();
         userId.remove();
         groupId.remove();
         setting.remove();
     }
 
-    // =================== 应用方法 ===================
+    // =================== 应用工具方法 ===================
 
     public static String getChatId() {
-        if (getIsPrivate()) return "Private_" + getUserId();
-        ChatScope scope = getSetting().getChatScope();
-        return scope + "_" + (scope == ChatScope.PERSONAL
+        if (getScope() == EventScope.PRIVATE) return "Private_" + getUserId();
+        ChatScope chatScope = getSetting().getChatScope();
+        return chatScope + "_" + (chatScope == ChatScope.PERSONAL
                 ? getUserId() : getGroupId());
     }
 }