+++ src/main/java/com/zincoid/nullbot/core/model/bot/event/CommandEvent.java
@@ -6,14 +6,16 @@ import com.mikuac.shiro.dto.event.message.PrivateMessageEvent;
 import com.mikuac.shiro.dto.event.notice.GroupMsgDeleteNoticeEvent;
 import com.mikuac.shiro.dto.event.notice.PokeNoticeEvent;
 import com.mikuac.shiro.enums.MsgTypeEnum;
+import com.zincoid.nullbot.core.context.BotCtx;
 import com.zincoid.nullbot.core.enums.EventScope;
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
-import lombok.Data;
+import lombok.Getter;
 
 import java.util.List;
 
-@Data
-@AllArgsConstructor(staticName = "of")
+@Getter
+@AllArgsConstructor(staticName = "of", access = AccessLevel.PRIVATE)
 public class CommandEvent<T extends Event> {
 
     private final T event;
@@ -22,7 +24,19 @@ public class CommandEvent<T extends Event> {
     private final boolean authRequired;
     private final boolean rateLimit;
 
-    // =================== 静态工厂方法 ===================
+    private final EventScope eventScope;
+    private final Long userId;
+    private final Long groupId;
+
+    // =================== 全参工厂方法 ===================
+
+    public static <T extends Event> CommandEvent<T> of(T event, String commandType,
+            List<String> commandParameters, boolean authRequired, boolean rateLimit) {
+        return CommandEvent.of(event, commandType, commandParameters, authRequired, rateLimit,
+                BotCtx.getScope(), BotCtx.getUserId(), BotCtx.getGroupId());
+    }
+
+    // =================== 便捷工厂方法 ===================
 
     public static CommandEvent<GroupMessageEvent> of(GroupMessageEvent event) {
         int i = event.getArrayMsg().getFirst().getType() == MsgTypeEnum.reply ? 1 : 0;
@@ -43,31 +57,5 @@ public class CommandEvent<T extends Event> {
 
     public static CommandEvent<GroupMsgDeleteNoticeEvent> of(GroupMsgDeleteNoticeEvent event) {
         return CommandEvent.of(event, "RecallReact", List.of(), false, false);
-    }
-
-    // =================== 信息获取方法 ===================
-
-    public EventScope getEventScope() {
-        if (event instanceof GroupMessageEvent) return EventScope.GROUP;
-        if (event instanceof PokeNoticeEvent e) return e.getGroupId() == null ? EventScope.PRIVATE : EventScope.GROUP;
-        if (event instanceof GroupMsgDeleteNoticeEvent) return EventScope.GROUP;
-        if (event instanceof PrivateMessageEvent) return EventScope.PRIVATE;
-        return EventScope.UNKNOWN;
-    }
-
-    public Long getGroupId() {
-        if (event instanceof GroupMessageEvent e) return e.getGroupId();
-        if (event instanceof PokeNoticeEvent e) return e.getGroupId() == null ? 0L : e.getGroupId();
-        if (event instanceof GroupMsgDeleteNoticeEvent e) return e.getGroupId();
-        if (event instanceof PrivateMessageEvent) return 0L;
-        throw new IllegalArgumentException("未知事件类型");
-    }
-
-    public Long getUserId() {
-        if (event instanceof GroupMessageEvent e) return e.getUserId();
-        if (event instanceof PokeNoticeEvent e) return e.getUserId();
-        if (event instanceof GroupMsgDeleteNoticeEvent e) return e.getUserId();
-        if (event instanceof PrivateMessageEvent e) return e.getUserId();
-        throw new IllegalArgumentException("未知事件类型");
     }
 }